backend/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
petl==1.7.4
requests==2.26.0
marshmallow==3.14.1
python-dateutil==2.8.2
parameterized==0.8.1
Pillow==9.0.0
black
easy-thumbnails==2.8.1
django-cors-headers==3.11.0
ipdb
